Update issue #216 Made some more changes to the Message API to allow easier manipulation of MetaData (which is now immutable)

+++ core/src/test/java/org/axonframework/domain/JavaSerializationTest.java
@@ -120,6 +120,6 @@ public class JavaSerializationTest {
 
         unmarshalled.doSomething();
 
-        assertEquals((Long) 1L, unmarshalled.getUncommittedEvents().next().getSequenceNumber());
+        assertEquals(1L, unmarshalled.getUncommittedEvents().next().getSequenceNumber());
     }
 }
