+++ integrationtests/src/test/java/org/axonframework/integrationtests/jpa/AbstractAnnotatedAggregateRoot_PersistenceTest.java
@@ -60,6 +60,6 @@ public class AbstractAnnotatedAggregateRoot_PersistenceTest {
         reloaded.doSomething();
 
         assertEquals(3, reloaded.getInvocationCount());
-        assertEquals((Long) 2L, reloaded.getUncommittedEvents().next().getSequenceNumber());
+        assertEquals(2L, reloaded.getUncommittedEvents().next().getSequenceNumber());
     }
 }
